+++ openeo/extra/artifacts/__init__.py
@@ -1,0 +1,1 @@
+from openeo.extra.artifacts.artifact_helper import build_artifact_helper